+++ CHANGELOG.md
@@ -4,6 +4,11 @@
   Placeholder for the next version (at the beginning of the line):
   ## **WORK IN PROGRESS**
 -->
+
+## **WORK IN PROGRESS**
+
+* (theimo1221) Try reconnect unreachable Daikin Devices
+
 ## 0.1.42-alpha.2 (2022-07-09)
 
 * (theimo1221) Fix release Script